feat[persistence]: Convert to a multi-module project
