+++ persistence/model/src/main/java/model/Article.java
@@ -1,4 +1,4 @@
-package org.company.persistence.model;
+package model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Getter;
