+++ persistence/model/src/main/java/model/NamedEntity.java
@@ -1,4 +1,4 @@
-package org.company.persistence.model;
+package model;
 
 import lombok.AccessLevel;
 import lombok.Getter;
