+++ persistence/repository/src/main/java/repository/PersistenceApplication.java
@@ -1,13 +1,14 @@
-package org.company.persistence;
+package repository;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class PersistenceApplication {
-// TODO - https://www.baeldung.com/spring-data-derived-queries
+// TODO
+//  - https://www.baeldung.com/spring-data-derived-queries
+//  - https://www.baeldung.com/rest-api-search-language-spring-data-specifications
     public static void main(String[] args) {
         SpringApplication.run(PersistenceApplication.class, args);
     }
-
 }
