+++ persistence/repository/src/main/java/repository/util/JpaSpecificationUtil.java
@@ -1,4 +1,4 @@
-package org.company.persistence.util;
+package repository.util;
 
 import lombok.experimental.UtilityClass;
 import org.springframework.data.jpa.domain.Specification;
