--- persistence/src/test/java/org/company/persistence/PersistenceApplicationTests.java
@@ -1,13 +1,0 @@
-package org.company.persistence;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PersistenceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
